🚚RENAME: change segmentation module place
**MODIFY:**
- 🧰 `Makefile`: change deprency place
- 🚚`src/main.c`: change include path

**RENAME:**
- 🚚`src/load_Image/auto_roation*` -> `src/segmentation/auto_roation*`:
  match module structure
- 🚚`src/load_Image/segmentation*` -> `src/segmentation/segmentation*`:
  match module structure
