+++ src/segmentation/auto_rotation.c
